unwarningize on vc8.  someone(tm) should still fix the whole nethandler->bzfs dependency thing though.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@14063 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/ServerLink.cxx
@@ -197,7 +197,7 @@ ServerLink::ServerLink(const Address& serverAddress, int port) :
     goto done;
 
   // send out the connect header
-  ::send(query,BZ_CONNECT_HEADER,strlen(BZ_CONNECT_HEADER),0);
+  ::send(query,BZ_CONNECT_HEADER,(int)strlen(BZ_CONNECT_HEADER),0);
 
   // get server version and verify
 #if !defined(_WIN32)
